requirements.txt
coverage>=5.1
matplotlib>=3.2.1
Pillow>=7.1.2
pylint>=2.5.2
pytest>=5.4.2
torch>=1.7.0
torchvision>=0.6.0
tqdm>=4.46.0